pyproject.toml
[tool.poetry]
name = "test-hub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "test_hub"}]

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^0.22.2"
django = "4.2.4"
django-treebeard = "^4.7"
pyjwt = "^2.8.0"
mysqlclient = "^2.2.0"
loguru = "^0.7.0"
django-guardian = "^2.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
